wezterm: Fix input lag on Windows when using WebGpu
In version 20240127-113634-bbcac864, the default frontend was switched
from OpenGL to WebGpu.

On my Windows setup, this caused a noticable input lag. I eventually
found a GitHub issue that also reported the same issue [1].

Apparently when on WebGpu, it might accidentally select an integrated
gpu as a default, which will cause the input lag issue.

I opened my debug overlay and confirmed that my wezterm also selected an
integrated gpu.

    Debug Overlay
    wezterm version: 20240127-113634-bbcac864 x86_64-pc-windows-msvc
    Window Environment: Windows
    WebGPU: name=Intel(R) UHD Graphics 770, device_type=IntegratedGpu, backend=Vulkan, driver=Intel Corporation, driver_info=Intel driver, vendor=32902, device=42880
    Enter lua statements or expressions and hit Enter.
    Press ESC or CTRL-D to exit

Once I switched it to a discrete gpu, the problem disappeared. Hence,
we shall make it select a discrete gpu as the preferred gpu, which is
what we want anyway. The code for that is directly taken from [2].

[1] https://github.com/wez/wezterm/issues/4278
[2] https://github.com/wez/wezterm/issues/4278#issuecomment-1726978476